pyproject.toml
[project]
name = "ticketswap-tap-redshift"
version = "0.0.1"
description = "Singer tap for Redshift, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Tobias Cadee", email = "tobias.cadee@ticketswap.com" }]
keywords = [
    "ELT",
    "Redshift",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "Apache-2.0"
license-files = [ "LICENSE" ]
requires-python = ">=3.9"
dependencies = [
    "singer-sdk>=0.41.0",
    "redshift-connector>=2.1.5",
    "boto3>=1.37.19",
    "sqlalchemy-redshift>=0.8.14",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
s3 = [
    "s3fs~=2025.5.0",
]

[project.scripts]
# CLI declaration
tap-redshift = 'tap_redshift.tap:TapRedshift.cli'

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "pytest>=8",
    "pytest-github-actions-annotate-failures>=0.3",
    "singer-sdk[testing]",
]
[tool.hatch.build.targets.wheel]
packages = [
    "tap_redshift",
]

[tool.pytest.ini_options]
addopts = [
    "--durations=10",
]

[tool.mypy]
warn_unused_configs = true
plugins = "sqlmypy"

[tool.ruff.lint]
ignore = [
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = [
    "hatchling>=1,<2",
]
build-backend = "hatchling.build"

# This configuration can be used to customize tox tests as well as other test frameworks like flake8 and mypy
[tool.tox]
min_version = "4.22"
requires = [
    "tox>=4.22",
    "tox-uv",
]
env_list = [
    "py313",
    "py312",
    "py311",
    "py310",
    "py39",
]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
pass_env = [
    "GITHUB_*",
    "TAP_REDSHIFT_*",
]
dependency_groups = [ "test" ]
commands = [ [ "pytest", { replace = "posargs", default = [ "tests" ], extend = true } ] ]
